+++ ext/duckdb/duckdb.c
@@ -47,5 +47,6 @@ Init_duckdb_native(void) {
     rbduckdb_init_duckdb_function_info();
     rbduckdb_init_duckdb_vector();
     rbduckdb_init_duckdb_data_chunk();
+    rbduckdb_init_memory_helper();
     rbduckdb_init_duckdb_table_function();
 }