+++ ext/duckdb/memory_helper.h
@@ -1,0 +1,8 @@
+/* frozen_string_literal: true */
+
+#ifndef RUBY_DUCKDB_MEMORY_HELPER_H
+#define RUBY_DUCKDB_MEMORY_HELPER_H
+
+void rbduckdb_init_memory_helper(void);
+
+#endif /* RUBY_DUCKDB_MEMORY_HELPER_H */